+++ lib/synchronized_call.dart
@@ -3,3 +3,4 @@ library synchronized_call;
 export 'src/base_lock.dart';
 export 'src/sync_lock.dart';
 export 'src/serial_lock.dart';
+export 'src/extra/inclusive_lock.dart';