+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.1
+
+- Refactor name `Lock` to `CallLock`
+- Add a new lock `InclusiveLock` that support for executing Head-Tail bloc task
+
+## 1.0.2
+
+- Enhance `InclusiveLock` finish callback